API/python/requirements.txt
aiosmtpd==1.4.4.post2
anyio==3.6.2
atpublic==3.1.1
attrs==22.2.0
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
fastapi==0.89.1
Flask==2.2.2
Flask-Cors==3.0.10
grpcio==1.51.1
grpcio-tools==1.51.1
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.19.0
packaging==23.0
protobuf==4.21.12
pycparser==2.21
pydantic==1.10.4
PyMySQL==1.0.2
python-dotenv==0.21.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
Werkzeug==2.2.2
zipp==3.11.0
